Episode 24

+++ app/Http/Controllers/ThreadController.php
@@ -23,7 +23,7 @@ class ThreadController extends Controller
     {
         $threads = $this->getThreads($channel, $filters);
 
-        if( request()->wantsJson()) {
+        if (request()->wantsJson()) {
             return $threads;
         }
 
@@ -110,9 +110,11 @@ class ThreadController extends Controller
      */
     public function destroy(Channel $channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
-        if ( request()->wantsJson() ) {
+        if (request()->wantsJson()) {
             return response([], 204);
         }
